# e-com/openapi.yaml
openapi: 3.0.3
info:
  title: E-commerce Backend API
  description: |
    REST API for e-commerce platform prototype. 
    Supports product browsing with filtering by category, brand, and price range, 
    as well as order management functionality.
  version: 1.0.0
  contact:
    name: E-commerce Team
    email: developer@ecommerce.com
  license:
    name: MIT License
    url: https://choosealicense.com/licenses/mit/
servers:
  - url: http://localhost:8080
    description: Development server
paths:
  /api/products:
    get:
      tags:
        - Products
      summary: Get all products with optional filtering
      description: Retrieve products with optional filters by category, brand, and price range
      parameters:
        - name: categoryId
          in: query
          description: Filter by category ID
          required: false
          schema:
            type: integer
            format: int64
        - name: brand
          in: query
          description: Filter by brand name (case insensitive)
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          description: Minimum price filter
          required: false
          schema:
            type: number
            format: double
            minimum: 0
        - name: maxPrice
          in: query
          description: Maximum price filter
          required: false
          schema:
            type: number
            format: double
            minimum: 0
        - name: search
          in: query
          description: Search by product name
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of products with category information
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Internal server error
    
    post:
      tags:
        - Products
      summary: Create a new product
      description: Add a new product to the catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input or category not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Internal server error
  /api/products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve a specific product by its ID
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
        '500':
          description: Internal server error
    
    put:
      tags:
        - Products
      summary: Update a product
      description: Update an existing product
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid input or category not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: Product not found
        '500':
          description: Internal server error
    
    delete:
      tags:
        - Products
      summary: Delete a product
      description: Remove a product from the catalog
      parameters:
        - name: id
          in: path
          description: Product ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Product deleted successfully
        '404':
          description: Product not found
        '500':
          description: Internal server error
  /api/products/category/{categoryId}:
    get:
      tags:
        - Products
      summary: Get products by category
      description: Retrieve all products in a specific category. By default returns product summaries without category info (since category is known from the path). Use includeCategory=true to get full products with category details.
      parameters:
        - name: categoryId
          in: path
          description: Category ID
          required: true
          schema:
            type: integer
            format: int64
        - name: includeCategory
          in: query
          description: Whether to include category information in the response
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of products in the category
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/ProductSummary'
                    description: Product summaries when includeCategory=false (default)
                  - type: array
                    items:
                      $ref: '#/components/schemas/Product'
                    description: Full products with category when includeCategory=true
        '404':
          description: Category not found
        '500':
          description: Internal server error
...
components:
  schemas:
    CategorySummary:
      type: object
      description: Lightweight category representation without products
      properties:
        id:
          type: integer
          format: int64
          description: Category ID
        name:
          type: string
          description: Category name
        description:
          type: string
          description: Category description
      required:
        - id
        - name
...
    ProductSummary:
      type: object
      description: Lightweight product representation without category details
      properties:
        id:
          type: integer
          format: int64
          description: Product ID
        name:
          type: string
          description: Product name
        description:
          type: string
          description: Product description
        price:
          type: number
          format: double
          minimum: 0
          description: Product price
        brand:
          type: string
          description: Product brand
        imageUrl:
          type: string
          description: Product image URL
        stockQuantity:
          type: integer
          minimum: 0
          description: Available stock quantity
      required:
        - id
        - name
        - price
        - brand
        - stockQuantity
    Product:
      type: object
      description: Full product representation with category summary
      properties:
        id:
          type: integer
          format: int64
          description: Product ID
        name:
          type: string
          description: Product name
        description:
          type: string
          description: Product description
        price:
          type: number
          format: double
          minimum: 0
          description: Product price
        brand:
          type: string
          description: Product brand
        imageUrl:
          type: string
          description: Product image URL
        stockQuantity:
          type: integer
          minimum: 0
          description: Available stock quantity
        category:
          $ref: '#/components/schemas/CategorySummary'
      required:
        - id
        - name
        - price
        - brand
        - category
        - stockQuantity
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
          description: Product name
          minLength: 1
        description:
          type: string
          description: Product description
        price:
          type: number
          format: double
          minimum: 0.01
          description: Product price
        brand:
          type: string
          description: Product brand
          minLength: 1
        imageUrl:
          type: string
          description: Product image URL
        stockQuantity:
          type: integer
          minimum: 0
          description: Available stock quantity
          default: 0
        categoryId:
          type: integer
          format: int64
          description: ID of the category this product belongs to
      required:
        - name
        - price
        - brand
        - categoryId
...